requirements.txt
Flask
gunicorn
beautifulsoup4
requests
python-dateutil
pytz